YAYYY! I fixed the average speed issue
Averages for concurrent mode were really slow, but having a queue_empty
bool and no repeating the loop on a delay has fixed the issue. This
changed it from pulling to a listener architecture. It still does pull
the queue_empty variable, but this is faster than locking an arc RwLock of the queue because it needs write permission.