Run tests!
The good news is that the sync is running as expected and writing data correctly. The concurrent is also running and writing correctly. The bad news is that the average over a column is not any better than I though it would be after the PR to optimize it. This means essentially it all is back how it was at the start of the day.